reformatted views for readability
