+++ media/output.json
@@ -196,7 +196,7 @@
           "Class Number": "14553",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
           "Time": "6:00PM - 8:40PM",
-          "Days": "M",
+          "Days": "T",
           "Location": "Remote Learning",
           "Instructor": "Karen Sue Allen"
         }
@@ -311,8 +311,8 @@
         "001": {
           "Class Number": "15200",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "4:30PM - 7:10PM",
-          "Days": "W",
+          "Time": "10:30AM - 11:45AM",
+          "Days": "MW",
           "Location": "Peter Kiewit Institute 259",
           "Instructor": "Ann L Fruhling"
         }
@@ -436,24 +436,24 @@
         "001": {
           "Class Number": "17587",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "8:30AM - 9:20AM",
-          "Days": "R",
+          "Time": "3:00PM - 4:15PM",
+          "Days": "TR",
           "Location": "Peter Kiewit Institute 278",
           "Instructor": "Mahadevan Subramaniam"
         },
         "002": {
           "Class Number": "17588",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
-          "Days": "TR",
+          "Time": "9:30AM - 10:20AM",
+          "Days": "T",
           "Location": "Peter Kiewit Institute 278",
           "Instructor": "Mahadevan Subramaniam"
         },
         "003": {
           "Class Number": "17589",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "1:30PM - 2:20PM",
-          "Days": "T",
+          "Time": "10:30AM - 11:20AM",
+          "Days": "W",
           "Location": "Peter Kiewit Institute 278",
           "Instructor": "Mahadevan Subramaniam"
         }
@@ -497,8 +497,8 @@
         "001": {
           "Class Number": "14260",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 8:40PM",
-          "Days": "M",
+          "Time": "10:30AM - 11:45AM",
+          "Days": "TR",
           "Location": "Durham Science Center 254",
           "Instructor": "Dora Velcsov"
         }
@@ -588,7 +588,7 @@
         "001": {
           "Class Number": "15236",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 7:15PM",
+          "Time": "7:30PM - 8:45PM",
           "Days": "TR",
           "Location": "Peter Kiewit Institute 252",
           "Instructor": "Hesham Hassan Ali"
@@ -618,7 +618,7 @@
         "001": {
           "Class Number": "16070",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "7:30PM - 8:45PM",
+          "Time": "9:00AM - 10:15AM",
           "Days": "TR",
           "Location": "Durham Science Center 104",
           "Instructor": "Fabio Torres Vitor"
@@ -710,8 +710,8 @@
         "820": {
           "Class Number": "14755",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
-          "Days": "TR",
+          "Time": "6:00PM - 8:40PM",
+          "Days": "M",
           "Location": "Remote Learning",
           "Instructor": "Bill L Sousan"
         }
@@ -725,7 +725,7 @@
         "001": {
           "Class Number": "14389",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
+          "Time": "7:30PM - 8:45PM",
           "Days": "TR",
           "Location": "Peter Kiewit Institute 263",
           "Instructor": "Eric Lundy"
@@ -733,8 +733,8 @@
         "820": {
           "Class Number": "20688",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "7:30PM - 8:45PM",
-          "Days": "TR",
+          "Time": "6:00PM - 8:40PM",
+          "Days": "M",
           "Location": "Peter Kiewit Institute 263",
           "Instructor": "Eric Lundy"
         }
@@ -763,8 +763,8 @@
         "001": {
           "Class Number": "14065",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:30PM - 4:20PM",
-          "Days": "R",
+          "Time": "3:00PM - 4:15PM",
+          "Days": "TR",
           "Location": "Peter Kiewit Institute 274",
           "Instructor": "Jorge Fandinno Garcia"
         }
@@ -778,8 +778,8 @@
         "001": {
           "Class Number": "14726",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
-          "Days": "TR",
+          "Time": "3:30PM - 4:20PM",
+          "Days": "R",
           "Location": "Peter Kiewit Institute 276",
           "Instructor": "Xin Zhong"
         }
@@ -824,7 +824,7 @@
           "Class Number": "20110",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
           "Time": "1:30PM - 2:45PM",
-          "Days": "TR",
+          "Days": "MW",
           "Location": "Durham Science Center 211A",
           "Instructor": "Mahboub Baccouch"
         }
@@ -853,8 +853,8 @@
         "001": {
           "Class Number": "15948",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "1:30PM - 2:45PM",
-          "Days": "MW",
+          "Time": "6:00PM - 8:40PM",
+          "Days": "W",
           "Location": "Peter Kiewit Institute 160",
           "Instructor": "Harvey Pe Siy"
         }
