+++ scheduleFunctions/templates/dashboard.html
@@ -154,7 +154,7 @@
           <a
             href="#"
             class="btn"
-            onclick="fetch('/run-solver/').then(r => r.json()).then(data => alert(data.message));"
+            onclick="fetch('/run-identifier/').then(r => r.json()).then(data => alert(data.message));"
             >Parse Upload</a
           >
           <button class="btn" onclick="optimizeSchedule()">
