+++ scheduleFunctions/views/__init__.py
@@ -1,0 +1,17 @@
+from .dashboard_view import *
+from .download_optimized_file import *
+from .login import *
+from .optimize_schedule import *
+from .run_clingo_identifier import *
+from .run_script import *
+from .upload_json_file import *
+
+__all__ = [
+    "dashboard_view",
+    "download_optimized_file",
+    "login",
+    "optimize_schedule",
+    "run_clingo_identifier",
+    "run_script",
+    "upload_json_file",
+]
