+++ scheduleFunctions/views/dashboard_view.py
@@ -1,0 +1,9 @@
+from django.shortcuts import render, redirect
+
+
+def dashboard_view(request):
+    if "username" not in request.session:
+        return redirect("home")  # Redirect to login if not authenticated
+
+    # Render the dashboard.html template
+    return render(request, "dashboard.html", {"username": request.session["username"]})
